backend/requirements.txt
Flask==1.1.1
Flask-Cors==2.1.2
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.4
